Remove unused component. Refactor seed bank selector into a standalone reusable component.
